#132 calculating power from energy delta for any HTTP/Modbus meter using non-configurable averaging interval instead of "measurement interval" energy is polled only every 60s schedule right before being polled by SHM energy polling is prefered over power polling minPulseDuration can be configured

+++ src/main/java/de/avanux/smartapplianceenabler/http/HttpHandler.java
@@ -40,26 +40,26 @@ public class HttpHandler implements ApplianceIdConsumer {
         this.httpTransactionExecutor = httpTransactionExecutor;
     }
 
-    public float getFloatValue(ParentWithChild<HttpRead, HttpReadValue> read,
+    public double getDoubleValue(ParentWithChild<HttpRead, HttpReadValue> read,
                                ContentProtocolHandler contentProtocolHandler) {
         String protocolHandlerValue = getValue(read, contentProtocolHandler);
         if(protocolHandlerValue != null) {
             String valueExtractionRegex = read.child().getExtractionRegex();
             String extractedValue = RegexUtil.getMatchingGroup1(protocolHandlerValue, valueExtractionRegex);
             String parsableString = extractedValue.replace(',', '.');
-            Float value;
+            Double value;
             Double factorToValue = read.child().getFactorToValue();
             if(factorToValue != null) {
-                value = Double.valueOf(Double.parseDouble(parsableString) * factorToValue).floatValue();
+                value = Double.parseDouble(parsableString) * factorToValue;
             }
             else {
-                value = Double.valueOf(Double.parseDouble(parsableString)).floatValue();
+                value = Double.parseDouble(parsableString);
             }
             logger.debug("{}: value={} protocolHandlerValue={} valueExtractionRegex={} extractedValue={}",
                     applianceId, value, protocolHandlerValue, valueExtractionRegex, extractedValue);
             return value;
         }
-        return 0.0f;
+        return 0.0;
     }
 
     public boolean getBooleanValue(ParentWithChild<HttpRead, HttpReadValue> read,
